+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/common/BaseEntity.java
@@ -16,8 +16,8 @@ import java.time.LocalDateTime;
 @Data
 public class BaseEntity {
 
-    @RowId
-    private String id;
+//    @RowId
+//    private String id;
 
     @CreatedDate
     @JsonDeserialize(using = LocalDateTimeDeserializer.class)