+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/common/LocalDateDeserializer.java
@@ -1,0 +1,26 @@
+package com.isxcode.oxygen.flysql.common;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+
+import java.io.IOException;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+/**
+ * LocalDate Deserializer
+ *
+ * @author ispong
+ * @since 0.0.1
+ */
+public class LocalDateDeserializer extends JsonDeserializer<LocalDate> {
+
+    @Override
+    public LocalDate deserialize(JsonParser p, DeserializationContext ctxt) throws IOException, JsonProcessingException {
+
+        DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        return LocalDate.parse(p.getText(), dtf);
+    }
+}