+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/core/FlysqlBuilder.java
@@ -9,6 +9,7 @@ import com.isxcode.oxygen.flysql.properties.FlysqlProperties;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.jdbc.core.JdbcTemplate;
 
+import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Objects;
 
@@ -43,21 +44,32 @@ public class FlysqlBuilder {
         if (mongoTemplate != null) {
             return DataBaseType.MONGO;
         } else {
+            Connection connection = null;
             try {
-                String databaseName = Objects.requireNonNull(jdbcTemplate.getDataSource()).getConnection().getMetaData().getDatabaseProductName();
+                connection = Objects.requireNonNull(jdbcTemplate.getDataSource()).getConnection();
+                String databaseName = connection.getMetaData().getDatabaseProductName();
                 switch (databaseName) {
                     case FlysqlConstants.ORACLE_DB:
                         return DataBaseType.ORACLE;
                     case FlysqlConstants.H2_DB:
                         return DataBaseType.H2;
                     case FlysqlConstants.MYSQL_DB:
+                    default:
                         return DataBaseType.MYSQL;
                 }
             } catch (SQLException e) {
+                e.printStackTrace();
                 throw new FlysqlException("datasource link error");
+            } finally {
+                if (connection != null) {
+                    try {
+                        connection.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
+                }
             }
         }
-        throw new FlysqlException("datasource link error");
     }
 
     /**