+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/response/SuccessResponse.java
@@ -10,13 +10,15 @@ public @interface SuccessResponse {
 
     /**
      * default msg
+     *
      * @return default msg
      */
     String value() default "";
 
     /**
      * msg
+     *
      * @return msg
      */
-    String msg() default "";
+    String msg();
 }